add reward system adapter
